WARNING: REDESIGNED: Expose structs as defined in specs
After lot of thinking I decided to redesign the library to expose the internal structures and make user `rresResourceChunk` struct directly aligned with the rres specificatio.

Previous implementation just exposed some of the `rresResourceChunkInfo` properties and converted the `unsigned char type[4]` (FourCC) to a more user-friendly `unsigned int type`, but it added a level of confusion and miss-aligned between the exposed `rresResourceChunk` and the structure defined in the specs. Now both are aligned.

The reasoning for previous implementation: Just expose the minimal required properties for the user and keep the non-required ones private to the library implementation.

The reasoning for new implementation: Compression/encryption was moved to user library, so, several of the properties were already exposed. The fields `flags`, `nextOfsset` and `crc32` could also be useful for user at some point so, why not expose them? And definitely, rres is an open specification and `rres.h` implementation is open source, there is no need to hide information to the users! And also the alignment betweent the specs and what user gets. The only degradation is the `type`, the user now gets a `unsigned char type[4]` instead of a `unsigned int type`, but `rresGetDataType()` function has been added to help (a bit) on that line.