Possible redesign under consideration...
Instead of using a variable number of properties, I considered the option of using a fixed number of 8 properties. This option simplifies `rres` specs implementation for other languages, it also simplifies a bit C code.

It also looks quite organized to me to have a 32 bytes header for resources +  a fixed 32 bytes properties data for all the resources.

The drawback is that some resource types could require more than 8 properties but all the current types do not need more than 4.

It's still an option under consideration.